[Feat] 알고리즘 통계 계산 로직 변경 (#34)

+++ Koco/src/main/java/icet/koco/user/dto/UserCategoryStatProjection.java
@@ -3,5 +3,5 @@ package icet.koco.user.dto;
 public interface UserCategoryStatProjection {
     Long getCategoryId();
     String getCategoryName();
-    Double getCorrectRate();
+    Integer getCorrectRate();
 }
